rename plugins/fixed-menu-hack
